+++ examples/typescript/http-server-pool/express/src/main.ts
@@ -30,7 +30,7 @@ expressApp.all('/api/echo', (req: Request, res: Response) => {
 expressApp.get('/api/factorial/:number', (req: Request, res: Response) => {
   const { number } = req.params
   requestHandlerPool
-    .execute({ body: { number } }, 'factorial')
+    .execute({ body: { number: parseInt(number) } }, 'factorial')
     .then(response => {
       return res.send(response.body).end()
     })