+++ examples/typescript/websocket-server-pool/ws/src/types.ts
@@ -1,0 +1,21 @@
+export enum MessageType {
+  echo = 'echo',
+  factorial = 'factorial'
+}
+
+export interface MessagePayload<T = unknown> {
+  type: MessageType
+  data: T
+}
+
+export interface DataPayload {
+  number?: number
+}
+
+export interface WorkerData<T = unknown> {
+  data: T
+}
+
+export interface WorkerResponse<T = unknown> {
+  data: T
+}